+++ packages/hydrogen/src/foundation/Router/DefaultRoutes.tsx
@@ -61,7 +61,7 @@ export function createRoutesFromPages(
 
   const routes = Object.keys(pages)
     .map((key) => {
-      const path = key
+      let path = key
         .replace('./pages', '')
         .replace(/\.server\.(t|j)sx?$/, '')
         /**
@@ -80,6 +80,9 @@ export function createRoutesFromPages(
           /\[(?:[.]{3})?(\w+?)\]/g,
           (_match, param: string) => `:${param}`
         );
+
+      if (path.endsWith('/') && path !== '/')
+        path = path.substring(0, path.length - 1);
 
       /**
        * Catch-all routes [...handle].jsx don't need an exact match